Only ever poll our Tasks on the same thread
Make sure that the first time we're polled that the current thread is remembered
and from there one we ensure that we're always polled from exactly the same thread.

In theory a GMainContext can be first run from one thread and later from another
thread, but we allow spawning non-Send futures and must not ever use them from
any other thread.